move function to helper
