+++ typescript/src/schema/resource-database.ts
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 import * as plugin from '../../../pbjs-genfiles/plugin';
+import {getResourceNameByPattern} from '../util';
 
 export interface ResourceDescriptor
   extends plugin.google.api.IResourceDescriptor {
@@ -90,7 +91,7 @@ export class ResourceDatabase {
         if (params.length === 0) {
           continue;
         }
-        const name = this.getName(pattern);
+        const name = getResourceNameByPattern(pattern);
         let resourceDescriptor: ResourceDescriptor = {
           name,
           params,
@@ -183,37 +184,6 @@ export class ResourceDatabase {
     let params = pattern.match(/{[a-zA-Z_]+(?:=.*?)?}/g) || [];
     params = params.map(p => p.replace(/{([a-zA-Z_]+).*/, '$1'));
     return params;
-  }
-
-  private getName(pattern: string): string {
-    const elements = pattern.split('/');
-    const name = [];
-    // Multi pattern like: `projects/{project}/cmekSettings`, we need to append `cmekSettings` to the name.
-    // Or it will be duplicate with `project/{project}`
-    // Iterate the elements, if it comes in pairs: user/{userId}, we take `userId` as part of the name.
-    // if it comes as `profile` with no following `/{profile_id}`, we take `profile` as part of the name.
-    // So for pattern: `user/{user_id}/profile/blurbs/{blurb_id}`, name will be `userId_profile_blurbId`
-    while (elements.length > 0) {
-      const eleName = elements.shift();
-      if (elements.length === 0) {
-        name.push(eleName);
-        break;
-      } else {
-        const nextEle = elements[0];
-        if (nextEle.match(/{[a-zA-Z_]+(?:=.*?)?}/g)) {
-          elements.shift();
-          name.push(
-            nextEle.substring(
-              1,
-              nextEle.includes('=') ? nextEle.indexOf('=') : nextEle.length - 1
-            )
-          );
-        } else {
-          name.push(eleName);
-        }
-      }
-    }
-    return name.join('_');
   }
 
   private getResourceDescriptor(
